+++ packages/database/src/repositories/knowledge/index.test.ts
@@ -3,7 +3,7 @@ import { FilesTabs } from '@lobechat/types';
 import { beforeEach, describe, expect, it } from 'vitest';
 
 import { getTestDB } from '../../models/__tests__/_util';
-import { NewDocument, documents } from '../../schemas/document';
+import { NewDocument, documents } from '../../schemas/file';
 import { NewFile, files } from '../../schemas/file';
 import { users } from '../../schemas/user';
 import { LobeChatDatabase } from '../../type';
