+++ packages/database/src/schemas/index.ts
@@ -3,7 +3,6 @@ export * from './aiInfra';
 export * from './apiKey';
 export * from './asyncTask';
 export * from './chatGroup';
-export * from './document';
 export * from './file';
 export * from './generation';
 export * from './message';
