+++ packages/database/src/schemas/relations.ts
@@ -6,8 +6,7 @@ import { createdAt } from './_helpers';
 import { agents, agentsFiles, agentsKnowledgeBases } from './agent';
 import { asyncTasks } from './asyncTask';
 import { chatGroups, chatGroupsAgents } from './chatGroup';
-import { documentChunks, documents } from './document';
-import { files, knowledgeBases } from './file';
+import { documentChunks, documents, files, knowledgeBases } from './file';
 import { generationBatches, generationTopics, generations } from './generation';
 import { messageGroups, messages, messagesFiles } from './message';
 import { chunks, unstructuredChunks } from './rag';
